feat: implement Google OAuth2 registration and token validation, add utility classes for email and phone number handling
