+++ backend/blocking/services/user-service/src/main/java/com/ticket4u/controller/AuthController.java
@@ -6,6 +6,7 @@ import com.ticket4u.dto.auth.*;
 import com.ticket4u.dto.auth.internal.LoginResult;
 import com.ticket4u.service.AuthService;
 import com.ticket4u.util.CookieUtil;
+import com.ticket4u.validation.GoogleTokenValidator;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -22,6 +23,7 @@ public class AuthController {
 
     private final CookieUtil cookieUtil;
     private final AuthService authService;
+    private final GoogleTokenValidator googleTokenValidator;
 
     /**
      * POST /api/auth/login
@@ -72,8 +74,14 @@ public class AuthController {
     public ResponseEntity<RegisterResponse> registerWithGoogle(
             @Valid @RequestBody OAuth2RegisterRequest oAuth2RegisterRequest
     ) {
-        log.debug("Google register request received");
-        RegisterResponse response = authService.registerWithGoogle(oAuth2RegisterRequest);
+        log.debug("Google register roAquest received");
+        
+        // Verify token and extract user info
+        GoogleIdToken idToken = new GoogleIdToken(oAuth2RegisterRequest.idToken());
+        GoogleUserInfo userInfo = googleTokenValidator.verifyAndExtract(idToken);
+        
+        // Register with verified user info
+        RegisterResponse response = authService.registerWithGoogle(userInfo);
         return ResponseEntity.ok(response);
     }
 }
