+++ backend/blocking/services/user-service/src/main/java/com/ticket4u/dto/auth/GoogleIdToken.java
@@ -1,0 +1,4 @@
+package com.ticket4u.dto.auth;
+
+public record GoogleIdToken(String value) {
+}
