+++ backend/blocking/services/user-service/src/main/java/com/ticket4u/dto/auth/GoogleUserInfo.java
@@ -1,0 +1,8 @@
+package com.ticket4u.dto.auth;
+
+public record GoogleUserInfo(
+        String email,
+        String name,
+        String picture
+) {
+}
